refactor(studio): split bulk easing helpers
